requirements.txt
Flask==3.0.3
yagmail==0.15.293
PyYAML==6.0.1
